examples/starter-python/requirements.txt
renowide-agent-sdk>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
PyYAML>=6.0.1
httpx>=0.27.0
